apps/api/pyproject.toml
[tool.poetry]
name = "slidedeck-ai-api"
version = "0.1.0"
description = "Backend for SlideDeck AI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-pptx = "^0.6.23"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.3"
google-generativeai = "^0.8.0"
openai = "^1.12.0"
anthropic = "^0.39.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
